# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "血癌照護管理平台",
        "description": "\n<h3>Application Interface Documentation for Blood Cancer Care Management Platform.</h3>\n<h4><a href='/'>Home Page</a></h4>\n                        ",
        "contact": {
            "name": "High Performance Distributed Systems Lab",
            "email": "nkust.ee.hpds@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/medicine/account/{account}": {
            "get": {
                "tags": [
                    "施打藥物及劑量紀錄"
                ],
                "summary": "取得施打藥物及劑量紀錄",
                "description": "取得施打藥物及劑量紀錄",
                "operationId": "ac6ecb5edd4fb6b349ff858b2ccb15bb",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "個案帳號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "description": "施打藥物及劑量紀錄編號",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "case_id": {
                                                    "description": "個案編號",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "date": {
                                                    "description": "施打日期",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "course": {
                                                    "description": "週期",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "type": {
                                                    "description": "藥物種類",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "dose": {
                                                    "description": "劑量",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reports/account/{account}": {
            "get": {
                "tags": [
                    "報告個管師紀錄"
                ],
                "summary": "取得報告個管師紀錄",
                "description": "取得報告個管師紀錄",
                "operationId": "037cb50b5567f396b8963289edb3059e",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "個案帳號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "description": "報告各管師編號",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "case_id": {
                                                    "description": "個案編號",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/report"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reports": {
            "post": {
                "tags": [
                    "報告個管師紀錄"
                ],
                "summary": "新增報告個管師紀錄",
                "description": "新增報告個管師紀錄",
                "operationId": "e54f4d802914b8df012b3d9bc9e4e2d1",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "account"
                                        ],
                                        "properties": {
                                            "account": {
                                                "description": "個案帳號",
                                                "type": "string",
                                                "example": "user1"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/report"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "description": "報告個管師編號",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "case_id": {
                                                    "description": "個案編號",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/report"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reports/{id}": {
            "delete": {
                "tags": [
                    "報告個管師紀錄"
                ],
                "summary": "刪除報告個管師紀錄",
                "description": "刪除報告個管師紀錄",
                "operationId": "a0564266d1760daf36eb7397dc9cbc82",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "報告個管師紀錄編號",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "patch": {
                "tags": [
                    "報告個管師紀錄"
                ],
                "summary": "更新報告個管師紀錄",
                "description": "更新報告個管師紀錄",
                "operationId": "a8e9ce28ae435d96c8e95999c4c88eef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "報告個管師紀錄編號",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/report"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "description": "報告個管師編號",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "case_id": {
                                                    "description": "個案編號",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/report"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "report": {
                "required": [
                    "date",
                    "physical_strength",
                    "symptom",
                    "hospital"
                ],
                "properties": {
                    "date": {
                        "description": "回報日期",
                        "type": "string",
                        "format": "date",
                        "example": "2022-03-10"
                    },
                    "physical_strength": {
                        "description": "體力狀況",
                        "type": "string",
                        "example": "很好"
                    },
                    "symptom": {
                        "description": "症狀",
                        "type": "string",
                        "example": "皮疹"
                    },
                    "hospital": {
                        "description": "固定回診醫院",
                        "type": "string",
                        "example": "高醫"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
    "security": [
        {
            "basicAuth": []
        }
    ],
...
}